Move `AsFd`'s lifetime parameter back into the function.
Based on usage feedback, move `AsFd`'s lifetime parameter back into the
function. This effectively reverts 910191b0.

This way is closer to how `AsRawFd` works, so it's more familiar, it's
simpler to write, and it avoids complex lifetime problems such as the
following:

```
struct BorrowedFd<'a> {
    r: &'a (),
}

trait AsFd<'a> {
    fn as_fd(self) -> BorrowedFd<'a>;
}

struct Bar<T> {
    t: T,
}

impl<T> Bar<T> {
    fn new<'a>(t: T)
    where
        &'a T: AsFd<'a>
    {
        use_it(&t);
        consume_it(t);
    }
}

fn use_it<'a, T>(t: &'a T) where &'a T: AsFd<'a>{
    let _ = t.as_fd();
}
fn consume_it<T>(_t: T) {
}
```

There doesn't seem to be a way to declare the `new` function
that satisfies the lifetime constraints. Generic parameters on the
`where` clause are a tempting idea, but rustc says that syntax is
reserved for future use.