Change `as_fd` from `&self` to `self`.
Change from

```rust
trait AsFd {
    fn as_fd(&self) -> BorrowedFd<'_>;
}
```

to

```rust
trait AsFd<'a> {
    fn as_fd(self) -> BorrowedFd<'a>;
}
```

and similar for `AsHandle`, `AsSocket`, `AsFilelike`, and
`AsSocketlike`. Instead of implementing `AsFd` for `OwnedFd` and other
owning types, it's now implemented for `&'a OwnedFd` and similar
references to other owning types.

This means somewhat more typing for implementors of the traits, but it
means fewer `&`s for users of APIs using the traits, and it means fewer
"reference to reference" situations.