Disable async-std and other dependencies on WASI. (#44)
* Disable async-std and other dependencies on WASI.

Move the async-std, tokio, socket2, and mio dependencies into a
`not(target_os = "wasi")` dependency section. This fixes the problem
that the default features in async-std, which we need, depend on
socket2, which doesn't work on WASI yet.

The tradeoff here is that WASI users won't be able to make use of
io-lifetimes' impls for these third-party crate types. But we are
working to add these traits upstream, which will eventually make
this unnecessary.

* Add testing on Rust 1.48.

* Add an MSRV note to README.md.

* Disable deprecation warnings in impls_std.rs.